Interactive mode for dodo
You must specify a flag `-i` for interactive mode

Note that stdin is only connected if you pipe text in, or you
specify the ``-i/--interactive`` flag.

This works around a bug in sshd, which causes all ssh connections
(including the sshfs mount) to hang, if reading from stdin while using
the master socket.
The downside is that interactive mode is slightly slower,
as it can't use the ssh master socket.